requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.1
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langsmith>=0.0.77
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pypdf>=3.0.0
jupyter>=1.0.0

# Meeting Briefing Agent
openai>=1.0.0

# Harmonic API integration
requests>=2.28.0

# Async HTTP (for batch Swarm API calls)
aiohttp>=3.8.0

# Tavily API (website intelligence)
tavily-python>=0.5.0

# Parallel Search API (news research)
parallel-web>=0.4.0

# Observability (optional - only needed if LANGSMITH_TRACING=true)
# langsmith is already included above

# API Server (for Lovable frontend)
fastapi>=0.100.0
uvicorn>=0.23.0

# Authentication (Phase 3.1 - Clerk JWT verification)
pyjwt[crypto]>=2.0.0
httpx>=0.24.0

# Output validation
pydantic>=2.0.0

# Testing
pytest>=7.0.0

# Supabase (PostgreSQL backend)
supabase>=2.0.0

# Pin protobuf to avoid Streamlit/grpc conflicts
protobuf>=4.0.0,<5.0.0
# Tavily integration enabled
# Force rebuild Wed Feb 25 14:56:23 PST 2026
